pyproject.toml
[tool.poetry]
name = "youtube-api-lab"
version = "0.1.0"
description = ""
authors = ["Cyrus.Tang <cyrus.tang@superoverflow.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
youtube-transcript-api = "^0.6.1"
openai = "^1.3.8"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
requests-mock = "^1.11.0"
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
